tracker/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
gspread>=5.12.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
